Move graphs to Graphs folder; create scaffolds
