+++ client/src/components/Graphs/Graph1.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Graph1() {
+  return (
+    <div>
+      
+    </div>
+  )
+}
