+++ client/src/components/Graphs/Graph2.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Graph2() {
+  return (
+    <div>
+      
+    </div>
+  )
+}
