+++ client/src/components/Graphs/Graph3.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Graph3() {
+  return (
+    <div>
+      
+    </div>
+  )
+}
